arc-camera: static_cast enum classes to int before logging
Log ostreams don't allow you to just log enum classes anymore
since they are not implicitly casted to int. Make that cast
explicit when logging.

BUG=b:37434548
TEST=emerge

Reviewed-on: https://chromium-review.googlesource.com/1086300
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>

+++ camera/common/v4l2_device/v4l2_subdevice.cc
@@ -63,7 +63,7 @@ int V4L2Subdevice::SetFormat(const struct v4l2_subdev_format& format) {
   VLOGF_ENTER();
   base::AutoLock l(state_lock_);
   if ((state_ != SubdevState::OPEN) && (state_ != SubdevState::CONFIGURED)) {
-    LOGF(ERROR) << "Invalid device state " << state_;
+    LOGF(ERROR) << "Invalid device state " << static_cast<int>(state_);
     return -EINVAL;
   }
 
@@ -97,7 +97,7 @@ int V4L2Subdevice::GetFormat(struct v4l2_subdev_format* format) {
   VLOGF_ENTER();
   base::AutoLock l(state_lock_);
   if ((state_ != SubdevState::OPEN) && (state_ != SubdevState::CONFIGURED)) {
-    LOGF(ERROR) << "Invalid device state " << state_;
+    LOGF(ERROR) << "Invalid device state " << static_cast<int>(state_);
     return -EINVAL;
   }
 
@@ -143,7 +143,7 @@ int V4L2Subdevice::SetSelection(const struct v4l2_subdev_selection& selection) {
   VLOGF_ENTER();
   base::AutoLock l(state_lock_);
   if ((state_ != SubdevState::OPEN) && (state_ != SubdevState::CONFIGURED)) {
-    LOGF(ERROR) << "Invalid device state " << state_;
+    LOGF(ERROR) << "Invalid device state " << static_cast<int>(state_);
     return -EINVAL;
   }
